# src/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Zadanie rekrutacyjne",
		"version": "0.1.9"
	},
	"host": "localhost:3000",
	"basePath": "",
	"definitions": {
		"id": {
			"properties": {
				"uuid": {
					"type": "string"
				}
			}
		}
	},
	"tags": [
		{
			"name": "TimeCounter",
			"description": "Aplikacja do śledzenia czasu pracy."
		}
	],
	"paths": {
		"/users/{userId}": {
			"parameters": [
				{
					"name": "userId",
					"in": "path",
					"required": true,
					"description": "User ID",
					"type": "string"
				}
			],
			"post": {
				"tags": ["TimeCounter"],
				"summary": "Create user with userID ",
				"responses": {
					"200": {
						"description": "Ok"
					},
					"400": {
						"description": "Fail"
					}
				}
			},
			"delete": {
				"tags": ["TimeCounter"],
				"summary": "Delete user with userID ",
				"responses": {
					"200": {
						"description": "Ok"
					},
					"400": {
						"description": "Fail"
					}
				}
			}
		},
		"/users/{userId}/{command}/{description}": {
			"parameters": [
				{
					"name": "userId",
					"in": "path",
					"required": true,
					"description": "User ID",
					"type": "string"
				},
				{
					"name": "command",
					"in": "path",
					"required": true,
					"description": "Commnad to start or stop timekeeping ",
					"schema": {
						"type": "string",
						"enum": ["start", "stop"]
					}
				},
				{
					"name": "description",
					"in": "path",
					"required": true,
					"description": "Short description of started task (ony for start command)",
					"type": "string"
				}
			],
			"post": {
				"tags": ["TimeCounter"],
				"summary": "Start or stop timekeeping for user with userID",
				"responses": {
					"200": {
						"description": "Ok"
					},
					"406": {
						"description": "Request rejected"
					},
					"400": {
						"description": "Fail"
					}
				}
			}
		},
		"/users/{userId}/data": {
			"parameters": [
				{
					"name": "userId",
					"in": "path",
					"required": true,
					"description": "User ID",
					"type": "string"
				}
			],
			"get": {
				"tags": ["TimeCounter"],
				"summary": "Get data from user. Valid only if at least one start/stop sequence had place",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Activity_time"
								}
							}
						}
					},
					"400": {
						"description": "Fail"
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Activity_time": {
				"type": "object",
				"properties": {
					"date": {
						"type": "string",
						"description": "Date of user activity"
					},
					"time": {
						"type": "number",
						"description": "Time of activity per day in ms"
					}
				}
			}
		}
	}
}